requirement.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
frozenlist==1.8.0
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
magic-filter==1.0.12
mccabe==0.7.0
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyflakes==3.4.0
pyjwt==2.10.1
pyopenssl==25.3.0
python-dotenv==1.2.1
pyyaml==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
sqlparse==0.5.4
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
typing-inspection==0.4.2
ujson==5.11.0
uritemplate==4.2.0
yarl==1.22.0
zope-interface==8.1.1
